Move fixed QE domain test into lib
